+++ .idea/compiler.xml
@@ -6,8 +6,8 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="AAA" />
         <module name="Generic" />
+        <module name="AAA" />
         <module name="ComCastXfinity" />
       </profile>
     </annotationProcessing>
