+++ .idea/encodings.xml
@@ -1,11 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/AAA/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/AAA/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/ComCastXfinity/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/ComCastXfinity/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/Generic/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/Generic/src/main/resources" charset="UTF-8" />
   </component>
 </project>
