+++ ComCastXfinity/src/test/java/testhome/TestHomePage.java
@@ -4,20 +4,22 @@ import common.WebAPI;
 import home.HomePage;
 import home.XfinityWebElement;
 import org.openqa.selenium.support.PageFactory;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import static home.HomePage.*;
+
 public class TestHomePage extends WebAPI {
-        public static HomePage homePage;
-        public static XfinityWebElement xfinityWebElement;
-        public static void init(){
-        homePage = PageFactory.initElements(driver,HomePage.class);
-        xfinityWebElement = PageFactory.initElements(driver,XfinityWebElement.class);
+    @BeforeMethod
+    public void ClassInIt(){
+        init();
+    }
+    @Test(enabled = true)
+    public static void homePage_get_Title_Test(){
+        homePage.homePage_get_Title();
     }
     @Test
-    public static void homePage_get_Title_Test(){
-        init();
-        homePage.homePage_get_Title();
-
+    public static void homePage_windowPopop_Test() throws InterruptedException {
+        homePage.homePage_windowPopop();
     }
-
 }
